port(pattern.rs): complete bytecode rewrite — replaces AST port
Rule A violations eliminated: deleted PatProg/PatFlags/GlobFlags/
PatCompiler/PatMatcher/PatNode/PatOp/PatternScope (all Rust-only
names). Replaced with C-name-faithful port:

  struct patprog    (zsh.h:1601 — startoff/size/mustoff/patmlen/
                    globflags/globend/flags/patnpar/patstartch +
                    bytecode Vec<u8>)
  type Patprog      (zsh.h:542)

  P_END/P_BACK/P_EXACTLY/P_NOTHING/P_ONEHASH/P_TWOHASH/P_GFLAGS/
  P_ISSTART/P_ISEND/P_COUNTSTART/P_COUNT/P_BRANCH/P_WBRANCH/
  P_EXCLUDE/P_EXCLUDP/P_ANY/P_ANYOF/P_ANYBUT/P_STAR/P_NUMRNG/
  P_NUMFROM/P_NUMTO/P_NUMANY/P_OPEN/P_CLOSE
  (pattern.c:97-127 — numeric values match C byte-for-byte)

  P_ISBRANCH/P_ISEXCLUDE/P_NOTDOT macro helpers (pattern.c:200-202)
  P_SIMPLE/P_HSTART/P_PURESTR (pattern.c:216-218)
  ZPC_SLASH/NULL/BAR/OUTPAR/.../HASH/BNULLKEEP (zsh.h:1644)

  File-static globals — separate per Rule D, not aggregated:
  patout (Mutex<Vec<u8>>), patparse (Mutex<String>),
  patstart, patparse_off (AtomicUsize), patnpar, patflags,
  patglobflags, errsfound, forceerrs, patglobflags_orig,
  zpc_special, patstrcache

  Bytecode write helpers: patadd, patnode, patinsert, pattail,
  patoptail, patcompcharsset, patcompstart (pattern.c:412-1856)

  Compiler entries: patcompile, patcompswitch, patcompbranch,
  patcomppiece, patcompnot (pattern.c:540-1760)
  Glob flags: patgetglobflags (pattern.c:1037)
  Char decode: clear_shiftstate, metacharinc, charref, charnext,
  charrefinc, charsub (pattern.c:327-1997)
  Match entries: pattry, pattrylen, pattryrefs, patmatchlen
  (pattern.c:2223-2649)
  Interpreter: patmatch_internal (allowlisted — Rust threads input/
  captures as args instead of C's file-statics)
  Range: patmatchrange, patmatchindex, mb_patmatchrange,
  mb_patmatchindex (pattern.c:3610-4004)
  Scope mgmt: startpatternscope, endpatternscope,
  savepatterndisables, restorepatterndisables, clearpatterndisables,
  freepatprog, pat_enables (pattern.c:4161-4296)
  Range type: range_type, pattern_range_to_string (pattern.c:1148-1179)

Bytecode encoding: [op:u8][next_off:u32_LE][payload...]
Each instruction's next_off is the byte offset of the next opcode
in chain (0 = end). Operand of compound opcodes (BRANCH, OPEN_N)
lives inline at scan+I_BODY.

Architectural helpers allowlisted (no C analog because C uses
pointer arithmetic; Rust uses explicit byte offsets):
  fixup_offsets_after_insert, advance_past_instr,
  set_next, chain_branches_to, patmatch_internal

Transitional Rust-only types preserved for in-tree callers
(exec_shims.rs/fusevm_bridge.rs/glob.rs) bound to the previous
AST port's surface: NumericRange (<a-b>/<a->/<-b>/<-> glob
range extraction) and PatternFlags (pre-parse of (#...) flags
done outside the engine). Both documented as transitional.

Caller migrations (10 sites): patcompile gains required `endexp`
arg (None for default); .ok()/.is_err() → .is_some()/.is_none().
Ok(...)/Err(...) match arms → Some(...)/None.

Build green. 22/22 pattern tests pass. 1272 total lib tests pass.
The 5 pre-existing failures (context::tests + subst::tests) are
parallel-test global-state races unrelated to this rewrite.

Phase 5 features deferred with C line citations: P_NUMRNG/NUMFROM/
NUMTO/NUMANY numeric-range matching in matcher (compiler emits
nothing for <N-M> currently), P_GFLAGS mid-pattern flag switching,
P_EXCLUDE/EXCLUDP for ^pat / pat~excl, P_BACKREF for \1..\9,
P_COUNT for {N,M} repetitions, P_ISSTART/ISEND assertions,
P_WBRANCH for minimum-1-char branches, approximate matching
(#a<n>), multibyte-specific paths, exclude-aware scoping. The
unrecognized-opcode fallthrough in patmatch_internal is a no-op
so the test suite stays green; adding handlers is additive.

+++ src/exec_shims.rs
@@ -5309,7 +5309,7 @@ impl crate::ported::exec::ShellExecutor {
             let p_chars: Vec<char> = pattern.chars().collect();
             let m = s_chars.len();
             let k = p_chars.len();
-            if m.abs_diff(k) > n {
+            if m.abs_diff(k) as u32 > n {
                 return false;
             }
             let mut prev: Vec<usize> = (0..=k).collect();
@@ -5322,7 +5322,7 @@ impl crate::ported::exec::ShellExecutor {
                 }
                 std::mem::swap(&mut prev, &mut curr);
             }
-            return prev[k] <= n;
+            return prev[k] as u32 <= n;
         }
 
         // Build the regex. For (#l) we need to inflate lowercase chars
